backend/requirements.txt
# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AWS SDK
boto3==1.34.0
botocore==1.34.0

# Database
sqlmodel==0.0.14
alembic==1.13.0
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Data Validation
pydantic==2.5.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-cognito==2024.5.1

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Logging & Monitoring
python-json-logger==2.0.7
structlog==23.2.0

# Environment Variables
python-dotenv==1.0.0

# CORS
fastapi-cors==0.0.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Background Tasks
celery==5.3.4
redis==5.0.1

# File Storage
aioboto3==12.1.0

# Date/Time
python-dateutil==2.8.2